minor bugfix

+++ src/dynein_simulate.cpp
@@ -121,12 +121,12 @@ void simulate(double runtime, double rand_seed, State init_state, double* init_p
 	  bool accept_step = false; 
 	  int attempts = 0;
 
-	  const int max_attempts = 1e10;
+	  const long long max_attempts = 1e10;
 	  while(!accept_step){
 	    if (attempts > max_attempts) {
-	      printf("Over %d attempts needed to avoid a NaN state in onebound, something must be wrong. Exiting.\n",
+	      printf("Over %lld attempts needed to avoid a NaN state in onebound, something must be wrong. Exiting.\n",
                      max_attempts);
-	      fprintf(stderr, "Over %d attempts needed to avoid a NaN state, something must be wrong. Exiting.\n",
+	      fprintf(stderr, "Over %lld attempts needed to avoid a NaN state, something must be wrong. Exiting.\n",
                       max_attempts);
 	      if (am_only_writing_on_crash) on_crash_write_movie_buffer();
 	      exit(1);
@@ -228,12 +228,12 @@ void simulate(double runtime, double rand_seed, State init_state, double* init_p
 	  bool accept_step = false;
 	  int attempts = 0;
 
-	  const int max_attempts = 1e10;
+	  const long long max_attempts = 1e10;
 	  while(!accept_step){
 	    if (attempts > max_attempts) {
-	      printf("Over %d attempts needed to avoid a NaN state in bothbound, something must be wrong. Exiting.\n",
+	      printf("Over %lld attempts needed to avoid a NaN state in bothbound, something must be wrong. Exiting.\n",
                      max_attempts);
-	      fprintf(stderr, "Over %d attempts needed to avoid a NaN state, something must be wrong. Exiting.\n",
+	      fprintf(stderr, "Over %lld attempts needed to avoid a NaN state, something must be wrong. Exiting.\n",
                       max_attempts);
 	      if (am_only_writing_on_crash) on_crash_write_movie_buffer();
 	      exit(1);
